requirements.txt
matplotlib==3.7.1
numpy==1.24.2
pandas==2.0.0
Pillow==9.4.0
Pillow==9.5.0
seaborn==0.12.2
